serde: Restructure serde module
