+++ zuri_nbt/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "zuri_nbt"
-version = "0.1.0"
+version = "0.2.0"
 edition = "2021"
 
 [dependencies]
